pyproject.toml
[tool.poetry]
name = "discord-shared-messages"
version = "0.1.0"
description = ""
authors = ["Henrique Gemignani Passos Lima <henrique@gemignani.org>"]

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^1.5.1"
sanic = ">=20"
Sanic-WTF = "^0.6.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
